feat: code style refacto

+++ src/engine/actions/handlers/CmdEject.cpp
@@ -5,56 +5,68 @@
 ** CmdEject.cpp
 */
 
-#include "CmdEject.hpp"
-
 #include <iostream>
 
-#include "../../entities/Player.hpp"
+#include "CmdEject.hpp"
 #include "../../World.hpp"
 #include "../../../ZappyServer.hpp"
 #include "../../../utils/Debug.hpp"
-#include "../../../utils/EventSystem.hpp"
 #include "../../../utils/EventRayDirectionInterpreter.hpp"
+#include "../../../utils/EventSystem.hpp"
+#include "../../entities/Player.hpp"
 
-void zappy::engine::cmd::CmdEject::cmdEject(std::weak_ptr<entities::Player> player, World& world, const std::string& args)
+void zappy::engine::cmd::CmdEject::cmdEject(std::weak_ptr<entities::Player> player, World& world,
+                                            const std::string& args)
 {
     const auto pusher = player.lock();
-    const auto& tile = world.getTileAt(pusher->getX(), pusher->getY());
+    const auto& tile = world.getTileAt(static_cast<int>(pusher->getX()), static_cast<int>(pusher->getY()));
 
     for (const auto egg : tile.getEggs())
         egg->crush();
 
     int dx = 0;
     int dy = 0;
-    switch (pusher->getDirection()) {
-        case Directions::NORTH: dx = 0; dy = -1; break;
-        case Directions::EAST: dx = 1; dy = 0; break;
-        case Directions::SOUTH: dx = 0; dy = 1; break;
-        case Directions::WEST: dx = -1; dy = 0; break;
+    switch (pusher->getDirection())
+    {
+        case Directions::NORTH: dx = 0;
+            dy = -1;
+            break;
+        case Directions::EAST: dx = 1;
+            dy = 0;
+            break;
+        case Directions::SOUTH: dx = 0;
+            dy = 1;
+            break;
+        case Directions::WEST: dx = -1;
+            dy = 0;
+            break;
     }
 
-    for (const auto victim : tile.getPlayers()) {
+    for (const auto victim : tile.getPlayers())
+    {
         if (victim->ID == pusher->ID)
             continue;
-        const auto& [newX, newY] = world.normalizeCoordinates(victim->getX() + dx, victim->getY() + dy);
+        const auto& [newX, newY] = world.normalizeCoordinates(static_cast<int>(victim->getX()) + dx, static_cast<int>(victim->getY()) + dy);
         victim->setPosition(newX, newY);
         world.getMainZappyServer().sendMessageToClient("eject: " +
-            std::to_string(utils::getRelativeDirection(
-                victim->getX(),
-                victim->getY(),
-                victim->getDirection(),
-                pusher->getX(),
-                pusher->getY(),
-                world.getWidth(),
-                world.getHeight(),
-                false)),
-            victim->ID);
+                                                       std::to_string(utils::getRelativeDirection(
+                                                           static_cast<int>(victim->getX()),
+                                                           static_cast<int>(victim->getY()),
+                                                           victim->getDirection(),
+                                                           static_cast<int>(pusher->getX()),
+                                                           static_cast<int>(pusher->getY()),
+                                                           world.getWidth(),
+                                                           world.getHeight(),
+                                                           false)),
+                                                       victim->ID);
         std::cout << debug::getTS() << "[TRACE] PLAYER " << pusher->ID << " PUSH PLAYER "
             << victim->ID << " TO CELL " << victim->getX() << ":" << victim->getY() << std::endl;
-	EventSystem::trigger("player_eject", world.getGraphicalClients(), world.getMainZappyServer().getConfig(), world, victim->ID);
+        EventSystem::trigger("player_eject", world.getGraphicalClients(), world.getMainZappyServer().getConfig(), world,
+                             victim->ID);
     }
 
-    for (const auto egg : tile.getEggs()) {
+    for (const auto egg : tile.getEggs())
+    {
         egg->crush();
         std::cout << debug::getTS() << "[TRACE] PLAYER " << pusher->ID <<
             " CRUSHED EGG " << egg->ID << " OF TEAM " << world.getTeamName(egg->getTeamID()) << std::endl;
