+++ src/engine/actions/handlers/CmdForward.cpp
@@ -22,22 +22,28 @@ namespace zappy::engine::cmd
 
         auto playerX = lockPlayer->getX();
         auto playerY = lockPlayer->getY();
-        world.getTileAt(playerX, playerY).removePlayer(lockPlayer);
+        world.getTileAt(static_cast<int>(playerX), static_cast<int>(playerY)).removePlayer(lockPlayer);
 
         switch (lockPlayer->getDirection())
         {
-            case Directions::NORTH: playerY -= 1; break;
-            case Directions::EAST: playerX += 1; break;
-            case Directions::SOUTH: playerY += 1; break;
-            case Directions::WEST: playerX -= 1; break;
+            case Directions::NORTH: playerY -= 1;
+                break;
+            case Directions::EAST: playerX += 1;
+                break;
+            case Directions::SOUTH: playerY += 1;
+                break;
+            case Directions::WEST: playerX -= 1;
+                break;
         }
 
-        const auto [newX, newY] = world.normalizeCoordinates(playerX, playerY);
+        const auto [newX, newY] = world.normalizeCoordinates(static_cast<int>(playerX), static_cast<int>(playerY));
         lockPlayer->setPosition(newX, newY);
-        world.getTileAt(lockPlayer->getX(), lockPlayer->getY()).addPlayer(lockPlayer);
+        world.getTileAt(static_cast<int>(lockPlayer->getX()), static_cast<int>(lockPlayer->getY())).addPlayer(lockPlayer);
 
-        std::cout << debug::getTS() << "[TRACE] Player " << lockPlayer->ID << " MOVED TO " << lockPlayer->getX() << ":" << lockPlayer->getY() << std::endl;
-	EventSystem::trigger("player_move", world.getGraphicalClients(), world.getMainZappyServer().getConfig(), world, lockPlayer->ID);
+        std::cout << debug::getTS() << "[TRACE] Player " << lockPlayer->ID << " MOVED TO " << lockPlayer->getX() << ":"
+            << lockPlayer->getY() << std::endl;
+        EventSystem::trigger("player_move", world.getGraphicalClients(), world.getMainZappyServer().getConfig(), world,
+                             lockPlayer->ID);
 
         world.getMainZappyServer().sendMessageToClient("ok", lockPlayer->ID);
     }
